move shared middleware
